requirements.txt
aiogram==3.21.0
python-dotenv==0.21.0
aiosqlite==0.19.0
